Implement cache invalidation + better local dev enviroment
