+++ supabase/config.toml
@@ -179,6 +179,9 @@ verify_jwt = false
 [functions.invalidate_cache]
 verify_jwt = false
 
+[functions.on_device_update]
+verify_jwt = false
+
 [functions.web_stats]
 verify_jwt = false
