+++ supabase/functions/on_channel_update/index.ts
@@ -3,6 +3,7 @@ import type { UpdatePayload } from '../_utils/supabase.ts'
 import { supabaseAdmin } from '../_utils/supabase.ts'
 import type { Database } from '../_utils/supabase.types.ts'
 import { getEnv, sendRes } from '../_utils/utils.ts'
+import { redisAppVersionInvalidate } from '../_utils/redis.ts'
 
 // Generate a v4 UUID. For this we use the browser standard `crypto.randomUUID`
 // function.
@@ -37,6 +38,17 @@ serve(async (event: Request) => {
       if (error)
         console.log('error', error)
     }
+
+    // Invalidate cache
+    if (!record.app_id) {
+      return sendRes({
+        status: 'error app_id',
+        error: 'Np app id included the request',
+      }, 500)
+    }
+
+    await redisAppVersionInvalidate(record.app_id)
+
     return sendRes()
   }
   catch (e) {
