+++ supabase/functions/on_device_update/.no_verify_jwt
@@ -1,0 +1,1 @@
+
